Migrating playSound function out of Livekit and WebRTC states. We fallback from playSound to playStream now.
