+++ play/src/front/Space/SpacePeerManager/WebRTCState.ts
@@ -83,10 +83,6 @@ export class WebRTCState implements ICommunicationState {
         return true;
     }
 
-    dispatchSound(url: URL): Promise<void> {
-        return this._peer.dispatchSound(url);
-    }
-
     dispatchStream(mediaStream: MediaStream): void {
         this._peer.dispatchStream(mediaStream);
     }
